Fix bags and setup webpack
