+++ .gitignore
@@ -1,3 +1,7 @@
+# Ignore bundler config.
+../public/bundle.js
+../public/bundle.js.map
+
 # Файлы и папки операционной системы
 .DS_Store
 Thumbs.db
