+++ public/index.html
@@ -1925,4 +1925,4 @@ Here is a card with filled data</textarea
   </body>
 </html>
 
-<script src = "../src/main.js"></script>
+<script src = "./bundle.js"></script>
